camera: Update mojo related code for libchrome urpev r456626.
So that it is almost compatible with current libchrome/libmojo and
newer ones.

BUG=b:37434548
TEST=Ran try.

Reviewed-on: https://chromium-review.googlesource.com/1297803
Commit-Ready: Wei Meng <wei.meng@intel.com>
Commit-Ready: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: Wei Meng <wei.meng@intel.com>
Tested-by: Hidehiko Abe <hidehiko@chromium.org>

+++ camera/camera3_test/camera3_frame_test.cc
@@ -273,7 +273,8 @@ Camera3FrameFixture::ConvertToImageAndRotate(BufferHandleUniquePtr buffer,
     return image;
   }
   if (format != ImageFormat::IMAGE_FORMAT_I420) {
-    LOG(ERROR) << "Do not support rotate image with format: " << format;
+    LOG(ERROR) << "Do not support rotate image with format: "
+               << static_cast<int>(format);
     return image;
   }
